RavenDB-24457 - Address PR Comments
